Test if Function Reproduces Stored Results

Call the function functionName with the arguments contained in testdata and compare the results with the results in testdata for identity.

test_function(
  functionName,
  testdata = loadArgs(functionName, file.path(tempdir(), "test")),
  dbg = TRUE
)

Arguments

functionName
Name of the function to test. It must be callable, i.e. either defined in the global environment or on the search path.

testdata
List of lists containing function arguments (in elemenet args) and results (in element result), just as returned by loadArgs. If no testdata are given, it is tried to load test data by calling loadArgs on functionName.

dbg
if TRUE (default) debug messages are shown

Value

TRUE If the function functionName is able to reproduce the same results as given in the result elements in testdata for all the argument combinations given in the args elements in testdata.

Examples

# Define a function using saveArgs() to save arguments and result
squareSum <- function(a, b) {
  result <- a * a + b * b
  saveArgs("squareSum", args = list(a = a, b = b), result = result)
  result
}

# Set global variable TESTMODE to "activate" saveArgs() in squareSum()
TESTMODE <- TRUE

# Call the function with different arguments
squareSum(1, 2)

#> [Set global variable TESTMODE to FALSE to prevent this message]
#> Save args to 'squareSum' (y, n)? 

#> [1] 5

squareSum(2, 3)

#> [Set global variable TESTMODE to FALSE to prevent this message]
#> Save args to 'squareSum' (y, n)? 

#> [1] 13

squareSum(-1, -2)

#> [Set global variable TESTMODE to FALSE to prevent this message]
#> Save args to 'squareSum' (y, n)? 

#> [1] 5

# The arguments and function results were saved here:
dir(file.path(tempdir(), "test"))

#> character(0)

# Write a new (wrong) version of the function
squareSum.new <- function(a, b) {
  a * a - b * b
}

# Check if it returns the same results
test_function("squareSum.new", loadArgs("squareSum"))

#> Error: No such file: 'test' in
#>   '/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpHveNBZ'.
#> Available files:
#>   'callr-env-76d2fa7660c'
#>   'downlit'
#>   'file76d249e7582'
#>   'file76d4ca1ce3'
#>   'file76da96c232'

# If no test data are given, loadArgs is called on the function to test,
# i.e. testing squareSum on the test data created by the same function will
# return TRUE if the function did not change in the meanwhile.
test_function("squareSum")

#> Error: No such file: 'test' in
#>   '/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpHveNBZ'.
#> Available files:
#>   'callr-env-76d2fa7660c'
#>   'downlit'
#>   'file76d249e7582'
#>   'file76d4ca1ce3'
#>   'file76da96c232'